// apps/api/src/modules/auth/domain/audit-event.enum.ts
export enum AuditEvent {
  USER_REGISTERED = 'USER_REGISTERED',
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGIN_FAILURE = 'LOGIN_FAILURE',
  LOGOUT = 'LOGOUT',
  TOKEN_REFRESH = 'TOKEN_REFRESH',
  PATIENT_CREATED = 'PATIENT_CREATED',
  PATIENT_LISTED = 'PATIENT_LISTED',
  PATIENT_SEARCHED = 'PATIENT_SEARCHED',
  PATIENT_VIEWED = 'PATIENT_VIEWED',
  PATIENT_UPDATED = 'PATIENT_UPDATED',
  PATIENT_DEACTIVATED = 'PATIENT_DEACTIVATED',
  PATIENT_EXPORTED = 'PATIENT_EXPORTED',
  MEDICAL_RECORD_CREATED = 'MEDICAL_RECORD_CREATED',
  MEDICAL_RECORD_VIEWED = 'MEDICAL_RECORD_VIEWED',
  MEDICAL_RECORD_LISTED = 'MEDICAL_RECORD_LISTED',
  MEDICAL_RECORD_UPDATED = 'MEDICAL_RECORD_UPDATED',
  MEDICAL_RECORD_SIGNED = 'MEDICAL_RECORD_SIGNED',
  MEDICAL_RECORD_ADDENDUM_CREATED = 'MEDICAL_RECORD_ADDENDUM_CREATED',
  CID10_SEARCHED = 'CID10_SEARCHED',
  NOTIFICATION_CREATED = 'NOTIFICATION_CREATED',
  NOTIFICATION_QUEUED = 'NOTIFICATION_QUEUED',
  NOTIFICATION_SENT = 'NOTIFICATION_SENT',
  NOTIFICATION_FAILED = 'NOTIFICATION_FAILED',
  NOTIFICATION_DASHBOARD_VIEWED = 'NOTIFICATION_DASHBOARD_VIEWED',
  NOTIFICATION_OPT_OUT_UPDATED = 'NOTIFICATION_OPT_OUT_UPDATED',
  DOCUMENT_UPLOAD_URL_CREATED = 'DOCUMENT_UPLOAD_URL_CREATED',
  DOCUMENT_UPLOAD_CONFIRMED = 'DOCUMENT_UPLOAD_CONFIRMED',
  DOCUMENT_VIEW_URL_CREATED = 'DOCUMENT_VIEW_URL_CREATED',
  DOCUMENT_LISTED = 'DOCUMENT_LISTED',
  DOCUMENT_SOFT_DELETED = 'DOCUMENT_SOFT_DELETED',
  CLINIC_CREATED = 'CLINIC_CREATED',
  TENANT_CONTEXT_RESOLVED = 'TENANT_CONTEXT_RESOLVED',
}